third_party/blink/web_tests/fast/forms/output/htmloutputelement-children-removed-expected.txt

Test that a live list is updated after the child of an HTMLOutputElement was removed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

The number of spans before removal is 2
The number of spans after removal is 1
PASS list.length == beforeLength - 1 is true
PASS successfullyParsed is true

TEST COMPLETE
